refactor: utils 폴더를 shared 내부로 이동시킨다.
